Abstracting out advance time helper functions and fixing a bug with advancing time after a checkpoint boundary
