+++ lib/agent0/agent0/interactive_fuzz/helpers/__init__.py
@@ -1,5 +1,6 @@
 """Shared functions for interactive fuzz testing."""
 
+from .advance_time import advance_time_after_checkpoint, advance_time_before_checkpoint
 from .close_random_trades import close_trades, permute_trade_events
 from .execute_random_trades import execute_random_trades
 from .fuzz_assertion_exception import FuzzAssertionException, fp_isclose
